implement h2_stream:terminate callback
The terminate callback makes it possible to notice when the
stream was taken down unexpectedly without receiving end_stream,
e.g. due to socket errors etc

In this case on_end_stream() is called also for termination. It will
result in a second eos being sent in case END_STREAM also was
received.