Fix interop tests
Fixes the interop tests so that they pass again. Tested against
both the go client and the grpcbox client. The issue was the calling
send_headers() lost the info that headers have been sent and so later
grpcbox ended up sending headers a second time. Storing the result
of send_headers() into the ctx prevents grpcbox from re-sending
the headers later.

This seems a bit hackish, not sure what the intended way of doing
this without tampering with the ctx is. Couldn't get it to work
using grpcbox_stream:add_headers() either.