Apply timeout to requests also while connecting
When the HTTP/2 connection is (re)connecting, no timeout applied
to further requests sent. They queue up in the gen_statem but
never expire. This commit makes the request return by setting
a timeout on the gen_statem:call function. This will not remove
the requests from the queue, but once that will not have a big
effect. Once the connection is established succeeds the queue will
empty quickly.