Fix error handling for streaming methods
Fixes an issue where grpcbox ignores errors that are returned from
the callback function of a streaming RPC. This happened both on
RPCs with streaming and non-streaming output, though the behaviour
was different: For non-streaming output, grpcbox always returned
a generic error, generated in `grpcbox_stream.erl` line 425 (on this
commit).
For streaming output, errors were completely ignored and the
stream was closed without an error indication to the client.

Also fixes a small issue where grpcbox passed incorrect ServerInfo
to streaming interceptors for RPCs with only streaming output.